+++ example/cbs_get_catalogs.R
@@ -1,3 +1,4 @@
+if (interactive()){
 catalogs <- cbs_get_catalogs()
 
 # Identifier of catalog can be used to query
@@ -6,3 +7,4 @@ print(catalogs$Identifier)
 
 ds_rivm <- cbs_get_datasets(catalog = "RIVM")
 ds_rivm[1:5, c("Identifier","ShortTitle")]
+}